# tests/yoke/http/golden/openapi.json
{
  "components": {
    "schemas": {
...
      "DurableEventInfo": {
        "additionalProperties": false,
        "properties": {
          "aggregateID": {
            "title": "Aggregateid",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "aggregateID",
          "seq"
        ],
        "title": "DurableEventInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PublicEvent"
            },
            "title": "Data",
            "type": "array"
          },
          "hasMore": {
            "title": "Hasmore",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "hasMore"
        ],
        "title": "HistoryResponse",
        "type": "object"
      },
...
      "LocationInfo": {
        "additionalProperties": false,
        "description": "Resolved workspace location.",
        "properties": {
          "directory": {
            "title": "Directory",
            "type": "string"
          }
        },
        "required": [
          "directory"
        ],
        "title": "LocationInfo",
        "type": "object"
      },
...
      "ModelCapabilities": {
        "additionalProperties": false,
        "properties": {
          "images": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          },
          "tools": {
            "default": true,
            "title": "Tools",
            "type": "boolean"
          }
        },
        "title": "ModelCapabilities",
        "type": "object"
      },
      "ModelInfo": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/ModelCapabilities"
          },
          "contextWindowTokens": {
            "title": "Contextwindowtokens",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "reasoningEfforts": {
            "items": {
              "type": "string"
            },
            "title": "Reasoningefforts",
            "type": "array"
          }
        },
        "required": [
          "id",
          "provider",
          "name",
          "reasoningEfforts",
          "capabilities",
          "contextWindowTokens"
        ],
        "title": "ModelInfo",
        "type": "object"
      },
      "ModelListResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            },
            "title": "Data",
            "type": "array"
          },
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          }
        },
        "required": [
          "location",
          "data"
        ],
        "title": "ModelListResponse",
        "type": "object"
      },
...
      "ProcessInfo": {
        "additionalProperties": false,
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "cwd": {
            "title": "Cwd",
            "type": "string"
          },
          "elapsedMs": {
            "title": "Elapsedms",
            "type": "integer"
          },
          "exitCode": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exitcode"
          },
          "output": {
            "$ref": "#/components/schemas/ProcessOutputInfo"
          },
          "pid": {
            "title": "Pid",
            "type": "integer"
          },
          "processID": {
            "title": "Processid",
            "type": "string"
          },
          "retention": {
            "const": "runtime",
            "default": "runtime",
            "title": "Retention",
            "type": "string"
          },
          "runtimeSessionID": {
            "title": "Runtimesessionid",
            "type": "integer"
          },
          "sessionID": {
            "title": "Sessionid",
            "type": "string"
          },
          "startedAt": {
            "title": "Startedat",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "exited",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "tty": {
            "title": "Tty",
            "type": "boolean"
          }
        },
        "required": [
          "processID",
          "sessionID",
          "runtimeSessionID",
          "pid",
          "command",
          "cwd",
          "tty",
          "status",
          "startedAt",
          "elapsedMs",
          "output"
        ],
        "title": "ProcessInfo",
        "type": "object"
      },
      "ProcessListResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProcessInfo"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "ProcessListResponse",
        "type": "object"
      },
      "ProcessOutputChunk": {
        "additionalProperties": false,
        "properties": {
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "stream": {
            "const": "combined",
            "default": "combined",
            "title": "Stream",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "text"
        ],
        "title": "ProcessOutputChunk",
        "type": "object"
      },
      "ProcessOutputCursor": {
        "additionalProperties": false,
        "properties": {
          "next": {
            "title": "Next",
            "type": "integer"
          },
          "truncatedBefore": {
            "title": "Truncatedbefore",
            "type": "integer"
          }
        },
        "required": [
          "next",
          "truncatedBefore"
        ],
        "title": "ProcessOutputCursor",
        "type": "object"
      },
      "ProcessOutputInfo": {
        "additionalProperties": false,
        "properties": {
          "latestSeq": {
            "title": "Latestseq",
            "type": "integer"
          },
          "originalBytes": {
            "title": "Originalbytes",
            "type": "integer"
          },
          "retainedBytes": {
            "title": "Retainedbytes",
            "type": "integer"
          },
          "tail": {
            "title": "Tail",
            "type": "string"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "tail",
          "originalBytes",
          "retainedBytes",
          "truncated",
          "latestSeq"
        ],
        "title": "ProcessOutputInfo",
        "type": "object"
      },
      "ProcessOutputResponse": {
        "additionalProperties": false,
        "properties": {
          "cursor": {
            "$ref": "#/components/schemas/ProcessOutputCursor"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProcessOutputChunk"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data",
          "cursor"
        ],
        "title": "ProcessOutputResponse",
        "type": "object"
      },
      "ProcessResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProcessInfo"
          }
        },
        "required": [
          "data"
        ],
        "title": "ProcessResponse",
        "type": "object"
      },
      "ProcessSignalRequest": {
        "additionalProperties": false,
        "properties": {
          "signal": {
            "enum": [
              "interrupt",
              "terminate"
            ],
            "title": "Signal",
            "type": "string"
          }
        },
        "required": [
          "signal"
        ],
        "title": "ProcessSignalRequest",
        "type": "object"
      },
      "ProcessStdinRequest": {
        "additionalProperties": false,
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ProcessStdinRequest",
        "type": "object"
      },
...
      "ProviderInfo": {
        "additionalProperties": false,
        "properties": {
          "currentModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentmodel"
          },
          "currentReasoningEffort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentreasoningeffort"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "id",
          "ready"
        ],
        "title": "ProviderInfo",
        "type": "object"
      },
      "ProviderListResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProviderInfo"
            },
            "title": "Data",
            "type": "array"
          },
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          }
        },
        "required": [
          "location",
          "data"
        ],
        "title": "ProviderListResponse",
        "type": "object"
      },
      "PublicEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "durable": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DurableEventInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "sessionID": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessionid"
          },
          "time": {
            "title": "Time",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "time",
          "data"
        ],
        "title": "PublicEvent",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Yoke HTTP API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/model": {
      "get": {
        "operationId": "listModels",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Directory"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Models"
      }
    },
    "/api/v1/process": {
      "get": {
        "operationId": "listProcesses",
        "parameters": [
          {
            "in": "query",
            "name": "sessionID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sessionid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "running",
                    "exited",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Processes"
      }
    },
    "/api/v1/process/{process_id}": {
      "get": {
        "operationId": "getProcess",
        "parameters": [
          {
            "in": "path",
            "name": "process_id",
            "required": true,
            "schema": {
              "title": "Process Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Process"
      }
    },
    "/api/v1/process/{process_id}/output": {
      "get": {
        "operationId": "getProcessOutput",
        "parameters": [
          {
            "in": "path",
            "name": "process_id",
            "required": true,
            "schema": {
              "title": "Process Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "afterSeq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Afterseq",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutputResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Process Output"
      }
    },
    "/api/v1/process/{process_id}/signal": {
      "post": {
        "operationId": "signalProcess",
        "parameters": [
          {
            "in": "path",
            "name": "process_id",
            "required": true,
            "schema": {
              "title": "Process Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessSignalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signal Process"
      }
    },
    "/api/v1/process/{process_id}/stdin": {
      "post": {
        "operationId": "writeProcessStdin",
        "parameters": [
          {
            "in": "path",
            "name": "process_id",
            "required": true,
            "schema": {
              "title": "Process Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessStdinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write Process Stdin"
      }
    },
    "/api/v1/provider": {
      "get": {
        "operationId": "listProviders",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Directory"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Providers"
      }
    },
...
    "/api/v1/session/{session_id}/history": {
      "get": {
        "operationId": "sessionHistory",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session History"
      }
    },
...
  }
}